+++ src/components/core/notify/DpNotifyContainer.vue
@@ -17,9 +17,14 @@
 <script>
 import { mapMutations, mapState } from 'vuex'
 import DpNotifyMessage from './DpNotifyMessage'
-import { hasOwnProp } from '@demos-europe/demosplan-utils'
-import { prefixClassMixin } from '@demos-europe/demosplan-utils'
+import { hasOwnProp } from '../../../utils'
+import { prefixClassMixin } from '../../../mixins'
 
+/**
+ * @deprecated
+ *
+ * import DpNotifyMessage directly and handle messages in outer component
+ **/
 export default {
   name: 'DpNotifyContainer',
 
@@ -98,7 +103,7 @@ export default {
         }
       }
 
-      document.addEventListener('visibilitychange', () => { this.isVisible = document.hidden })
+      document.addEventListener('visibilitychange', () => { this.isVisible = !document.hidden })
     },
 
     removeMessage (message) {