+++ src/components/index.js
@@ -1,62 +1,64 @@
+import DpAccordion from './DpAccordion/DpAccordion'
 import DpButton from './DpButton/DpButton'
+import DpButtonRow from './DpButtonRow/DpButtonRow'
+import DpCheckbox from './DpCheckbox/DpCheckbox'
+import DpCheckboxGroup from './DpCheckboxGroup/DpCheckboxGroup'
+import DpContextualHelp from './DpContextualHelp/DpContextualHelp'
+import DpDatepicker from './DpDatepicker/DpDatepicker'
+import DpDateRangePicker from './DpDateRangePicker/DpDateRangePicker'
+import DpDatetimePicker from './DpDatetimePicker/DpDatetimePicker'
 import DpDetails from './DpDetails/DpDetails'
+import DpFlyout from './DpFlyout/DpFlyout'
+import DpFormRow from './DpFormRow/DpFormRow'
 import DpIcon from './DpIcon/DpIcon'
+import DpInlineNotification from './DpInlineNotification/DpInlineNotification'
 import DpInput from './DpInput/DpInput'
 import DpLabel from './DpLabel/DpLabel'
 import DpLoading from './DpLoading/DpLoading'
+import DpMultiselect from './DpMultiselect/DpMultiselect'
+import DpPager from './DpPager/DpPager'
+import DpProgressBar from './DpProgressBar/DpProgressBar'
+import DpRadio from './DpRadio/DpRadio'
+import DpResettableInput from './DpResettableInput/DpResettableInput'
+import DpSearchField from './DpSearchField/DpSearchField'
+import DpSelect from './DpSelect/DpSelect'
+import DpSlidingPagination from './DpSlidingPagination/DpSlidingPagination'
+import DpSplitButton from './DpSplitButton/DpSplitButton'
+import DpTextArea from './DpTextArea/DpTextArea'
+import DpTimePicker from './DpTimePicker/DpTimePicker'
+import DpToggle from './DpToggle/DpToggle'
 import {
   dataTableSearch,
-  DpAccordion,
   DpAnonymizeText,
   DpAutocomplete,
   DpBulkEditHeader,
   DpButtonIcon,
-  DpButtonRow,
   DpCard,
   DpChangeStateAtDate,
-  DpCheckbox,
-  DpCheckboxGroup,
   DpColumnSelector,
-  DpContextualHelp,
   DpCopyPasteButton,
   DpDashboardTaskCard,
   DpDataTable,
   DpDataTableExtended,
-  DpDatepicker,
-  DpDateRangePicker,
-  DpDatetimePicker,
+  DpDraggable,
   DpEditableList,
   DpEditor,
-  DpFlyout,
-  DpFormRow,
   DpHeightLimit,
-  DpInlineNotification,
   DpModal,
-  DpMultiselect,
   DpNotifyContainer,
+  DpNotifyMessage,
   DpObscure,
-  DpPager,
-  DpProgressBar,
-  DpRadio,
   DpRegisterFlyout,
-  DpResettableInput,
-  DpSearchField,
-  DpSelect,
   DpSelectPageItemCount,
   DpSkeletonBox,
   DpSlidebar,
-  DpSlidingPagination,
-  DpSplitButton,
   DpStickyElement,
   DpSwitcher,
   DpTab,
   DpTableCard,
   DpTableCardListHeader,
   DpTabs,
-  DpTextArea,
   DpTextWrapper,
-  DpTimePicker,
-  DpToggle,
   DpToggleForm,
   DpTooltipIcon,
   DpTransitionExpand,
@@ -68,19 +70,19 @@ import {
 } from './core/'
 
 export {
+  DpAccordion,
   DpButton,
+  DpButtonRow,
   DpDetails,
   DpIcon,
   DpInput,
   DpLabel,
   DpLoading,
   dataTableSearch,
-  DpAccordion,
   DpAnonymizeText,
   DpAutocomplete,
   DpBulkEditHeader,
   DpButtonIcon,
-  DpButtonRow,
   DpCard,
   DpChangeStateAtDate,
   DpCheckbox,
@@ -94,6 +96,7 @@ export {
   DpDatepicker,
   DpDateRangePicker,
   DpDatetimePicker,
+  DpDraggable,
   DpEditableList,
   DpEditor,
   DpFlyout,
@@ -103,6 +106,7 @@ export {
   DpModal,
   DpMultiselect,
   DpNotifyContainer,
+  DpNotifyMessage,
   DpObscure,
   DpPager,
   DpProgressBar,