+++ src/mixins/prefixClassMixin/prefixClassMixin.js
@@ -1,9 +1,7 @@
 /**
- * This mixin can be to provide the prefixClass method and a prop to control when to prefix and when not.
- *
- * @deprecated USe prefixClassMixin from demosplan-utils instead
+ * This mixin can be used to provide the prefixClass method and a prop to control when to prefix and when not.
  */
-import { prefixClass } from '../../lib/'
+import { prefixClass } from '../../utils'
 
 export default {
   methods: {