+++ webpack.config.js
@@ -1,6 +1,12 @@
 const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const { VueLoaderPlugin } = require('vue-loader');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
+
+const bundleAnalyzer = new BundleAnalyzerPlugin({
+  analyzerMode: 'static',
+  reportFilename: resolve(`./bundle_analysis.html`)
+})
 
 const isProduction = process.env.NODE_ENV == 'production';
 
@@ -10,15 +16,41 @@ function resolve (dir) {
   return path.join(__dirname, dir)
 }
 
+const transpileNodeModules = [
+  'tiptap',
+  'tiptap-commands',
+  'tiptap-extensions',
+].map(module => resolve('node_modules/' + module))
+
 const config = {
   entry: resolve('./src/index.js'),
   output: {
     path: path.resolve(__dirname, 'dist'),
+    publicPath: '',
     filename: 'demosplan-ui.umd.js',
-    library: '@demos-europe/demosplan-ui',
+    library: {
+      name: '__demos_europe_demosplan_ui',
+      type: 'var'
+    },
+    libraryExport: 'default',
     libraryTarget: 'umd',
-    libraryExport: 'default'
+    clean: true
   },
+  externalsType: 'commonjs',
+  externals: [
+    '@braintree/sanitize-url',
+    /^@uppy\/.+$/,
+    'dayjs',
+    'dompurify',
+    'lscache',
+    'plyr',
+    'tippy.js',
+    'uuid',
+    'v-tooltip',
+    'vue-multiselect',
+    'vuedraggable',
+    'vuex'
+  ],
   resolve: {
     extensions: ['.js', '.vue'],
     symlinks: false
@@ -36,6 +68,7 @@ const config = {
       {
         test: /\.(js|jsx)$/i,
         exclude: /node_modules/,
+        include: transpileNodeModules,
         loader: 'babel-loader'
       },
       {
@@ -59,6 +92,10 @@ module.exports = () => {
     config.mode = 'production';
   } else {
     config.mode = 'development';
+  }
+
+  if (process.argv.includes('--analyze')) {
+    config.plugins.unshift(bundleAnalyzer)
   }
 
   return config;